[Services] Refactorized some namespace services
